Receive Transaction Data As Count Per Second

+++ pkg/monitor/types.go
@@ -1,10 +1,10 @@
 package monitor
 
 type Transaction struct {
-	ServiceId           string         `json:"serviceID,omitempty"`
-	EndpointsCounterMap map[string]int `json:"endpointCounter,omitempty"`
+	ServiceId           string             `json:"serviceID,omitempty"`
+	EndpointsCounterMap map[string]float64 `json:"endpointCounter,omitempty"`
 }
 
-func (this *Transaction) GetEndpointsCounterMap() map[string]int {
+func (this *Transaction) GetEndpointsCounterMap() map[string]float64 {
 	return this.EndpointsCounterMap
 }
